new ticket is 90% done and started adding repsonding
responding doesnt seem to want to work at the moment. I can not get data to be inserted into the database. the sql string seems to be working fine but it is not working. I can not thinking of any reason why it would not work at the moment. Going to table it for now and come back to it later.